Update libgit2 to ff8d635
https://github.com/libgit2/libgit2/compare/9bbc8f3...ff8d635

+++ LibGit2Sharp/Core/GitIndexEntry.cs
@@ -15,7 +15,7 @@ namespace LibGit2Sharp.Core
         public uint Mode;
         public uint Uid;
         public uint Gid;
-        public Int64 file_size;
+        public uint file_size;
         public GitOid Id;
         public ushort Flags;
         public ushort ExtendedFlags;
